\name{Reformat32}
\alias{Reformat32}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Reformat data}
\description{Reformat data from the 3-way table format (containing counts of the various alellic
  states) to the 2-way  table format (containing allelic  states)}

\usage{Reformat32(g)}
%- maybe also 'usage' for other objects documented here.
\arguments{  \item{g}{An array in  the 3-way table format } }
\value{A matrix in the 2-way  table format (containing allelic  states)}
\author{Casper Eriksen}
\keyword{Data format }

