\name{Sunder-package}
\alias{Sunder-package}
\alias{Sunder}
\docType{package}
\title{Inference and model selection for analysis of geographical
  genetic variation}
\description{MCMC inference and model selection by cross-validation for the analysis of geographical  genetic variation}
\details{
\tabular{ll}{
Package: \tab Sunder\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-20\cr
License: \tab What license is it under?\cr
}
Simulation, MCMC inference and model selection by cross-validation for the analysis
of geographical  genetic variation

}
\author{Filippo Botta and Gilles Guilllot

Maintainer: Filippo Botta  <filippo.botta@gmail.com>
}
\references{Cf. \code{citation(Sunder)}}

\keyword{ package }
%\seealso{}
%\examples{}
