\name{Sunder-package}
\alias{Sunder-package}
\alias{Sunder}
\docType{package}
\title{Inference and model selection for analysis of geographical
  genetic variation}
\description{Data simulation, inference and model selection by cross-validation for the analysis of geographical  genetic variation}

\author{Filippo Botta, Casper Eriksen and Gilles Guilllot
  Maintainer: Filippo Botta  <filippo.botta@gmail.com>
}
\references{Cf. \code{citation(Sunder)}}

\keyword{ package }

\details{
\tabular{ll}{
  Package: \tab Sunder\cr
  Type: \tab Package\cr
Version: \tab 0.0.4 \cr
Date: \tab Mon Jan 12 19:54:04 2015 \cr
License: \tab GPL \cr
}
Simulation, inference and model selection by cross-validation for the analysis of geographical  genetic variation
}
